+++ tests/testthat/test-env.R
@@ -385,3 +385,16 @@ test_that("env_bind_fns() uses as_function()", {
   env_bind_fns(current_env(), foo = ~2 + 3)
   expect_identical(foo, 5)
 })
+
+test_that("env_label() returns proper environment name", {
+  expect_identical(env_label(global_env()), "global")
+  expect_identical(env_label(empty_env()), "empty")
+  expect_identical(env_label(base_env()), "base")
+
+  expect_identical(env_label(pkg_env("rlang")), "package:rlang")
+  expect_identical(env_label(ns_imports_env("rlang")), "imports:rlang")
+  expect_identical(env_label(ns_env("rlang")), "namespace:rlang")
+
+  env <- structure(env(), name = "foobar")
+  expect_identical(env_label(env), "foobar")
+})